+++ custom/package-local.sh
@@ -41,8 +41,8 @@ cat > "$APP/Contents/Info.plist" <<'PLIST'
 <key>CFBundleDisplayName</key><string>YuE Studio</string>
 <key>CFBundleIdentifier</key><string>com.solution7.yuestudio.custom</string>
 <key>CFBundleExecutable</key><string>YuE Studio</string>
-<key>CFBundleVersion</key><string>20260917.2</string>
-<key>CFBundleShortVersionString</key><string>0.5.0</string>
+<key>CFBundleVersion</key><string>20260917.3</string>
+<key>CFBundleShortVersionString</key><string>0.5.1</string>
 <key>CFBundlePackageType</key><string>APPL</string>
 <key>CFBundleIconFile</key><string>AppIcon</string>
 <key>LSMinimumSystemVersion</key><string>14.0</string>